add f64/f32 features to kesko_physics
in order to switch between f32 and f64 for rapier the physic crate now reflect the same feature.
the cut is not super clear. some conversion are needed in other creates and maybe more creates should have this feature.

+++ crates/kesko_physics/src/gravity.rs
@@ -1,5 +1,5 @@
 use bevy::prelude::*;
-use rapier3d::dynamics::RigidBodySet;
+use crate::rapier_extern::rapier::dynamics::RigidBodySet;
 use crate::rigid_body::RigidBodyHandle;
 
 
@@ -22,8 +22,8 @@ impl Default for Gravity {
 
 #[derive(Component)]
 pub struct GravityScale {
-    pub val: f32,
-    pub initial_val: f32
+    pub val: f64,
+    pub initial_val: f64
 }
 
 impl Default for GravityScale {
@@ -33,7 +33,7 @@ impl Default for GravityScale {
 }
 
 impl GravityScale {
-    pub fn new(val: f32) -> Self {
+    pub fn new(val: f64) -> Self {
         Self { val, initial_val: val }
     }
